Extract wardley_mapping/ bounded context
Move WM-specific types (TourStop, TourManifest, TourManifestRepository),
WardleyProjectPresenter, RegisterTourUseCase, and tour DTOs into the
wardley_mapping/ package.

Introduces ProjectLookup protocol so the wardley_mapping BC can verify
project existence without importing from consulting/. ProjectRepository
already satisfies its shape, so DI passes it through with no adapter.

Part of #47.
