+++ bin/cli/dtos.py
@@ -1,12 +1,11 @@
 """Request and response DTOs — re-exports from consulting.dtos
-plus tour and site DTOs that remain in bin/cli.
+and wardley_mapping.dtos, plus site DTOs that remain in bin/cli.
 """
 
 from __future__ import annotations
 
 from pydantic import BaseModel, Field
 
-from bin.cli.wm_types import TourStop
 from consulting.dtos import (
     AddEngagementEntryRequest,
     AddEngagementEntryResponse,
@@ -35,6 +34,7 @@ from consulting.dtos import (
     UpdateProjectStatusRequest,
     UpdateProjectStatusResponse,
 )
+from wardley_mapping.dtos import RegisterTourRequest, RegisterTourResponse
 
 __all__ = [
     "AddEngagementEntryRequest",
@@ -68,31 +68,6 @@ __all__ = [
     "UpdateProjectStatusRequest",
     "UpdateProjectStatusResponse",
 ]
-
-
-# ---------------------------------------------------------------------------
-# RegisterTour — stays here until wardley_mapping BC is extracted
-# ---------------------------------------------------------------------------
-
-
-class RegisterTourRequest(BaseModel):
-    """Register or replace a presentation tour for a project."""
-
-    client: str = Field(description="Client slug.")
-    project_slug: str = Field(
-        description="Project slug.",
-        json_schema_extra={"cli_name": "project"},
-    )
-    name: str = Field(description="Tour name (e.g. investor).")
-    title: str = Field(description="Tour display title.")
-    stops: list[TourStop] = Field(description="JSON array of tour stops.")
-
-
-class RegisterTourResponse(BaseModel):
-    client: str
-    project_slug: str
-    name: str
-    stop_count: int
 
 
 # ---------------------------------------------------------------------------
